+++ local/testing/api-tests/tavern_helpers.py
@@ -111,14 +111,11 @@ def relationships_are_not_empty(response):
 
 
 def submodels_are_empty(response):
-    print(response.json().get("submodels"))
     print("Check if submodels are empty", len(response.json().get("submodels")))
     assert len(response.json().get("submodels")) == 0
 
 
 def submodels_are_not_empty(response):
-    print(response)
-    print(response.json().get("submodels"))
     print("Check if submodels are not empty", len(response.json().get("submodels")))
     assert len(response.json().get("submodels")) != 0
 
@@ -140,7 +137,7 @@ def submodelDescriptors_in_shells_are_empty(response):
     shells = response.json().get("shells")
     print("shells ", shells)
     for i in shells:
-        assert len(shells.get("submodelDescriptors")) == 0
+        assert len(i.get("submodelDescriptors")) == 0
 
 
 def aspects_in_job_parameter_are_empty(response):