+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added cache mechanism in DiscoveryFinderClientImpl for findDiscoveryEndpoints
 
 ## [4.3.0] - 2023-12-08
 ### Added